refactor(shared): extract id regex into separate module
