+++ apps/papra-server/src/modules/custom-properties/options/custom-properties-options.constants.ts
@@ -1,4 +1,4 @@
-import { createPrefixedIdRegex } from '../../shared/random/ids';
+import { createPrefixedIdRegex } from '../../shared/random/ids.constants.models';
 
 export const CUSTOM_PROPERTY_SELECT_OPTION_ID_PREFIX = 'cpso';
 export const CUSTOM_PROPERTY_SELECT_OPTION_ID_REGEX = createPrefixedIdRegex({ prefix: CUSTOM_PROPERTY_SELECT_OPTION_ID_PREFIX });
