+++ apps/papra-server/src/modules/tags/tags.constants.ts
@@ -1,4 +1,4 @@
-import { createPrefixedIdRegex } from '../shared/random/ids';
+import { createPrefixedIdRegex } from '../shared/random/ids.constants.models';
 
 export const TagColorRegex = /^#[0-9A-F]{6}$/;
